refactor(minecraft): extract instance list loading into ModInstanceList
