+++ Plain Craft Launcher 2/Modules/ModEvent.cs
@@ -187,12 +187,12 @@ namespace PCL
                     case EventType.启动游戏:
                         if (args[0] == "\\current")
                         {
-                            if (ModMinecraft.McMcInstanceSelected is null)
+                            if (ModInstanceList.McMcInstanceSelected is null)
                             {
                                 ModMain.Hint("请先选择一个 Minecraft 版本！", ModMain.HintType.Critical);
                                 return;
                             }
-                            args[0] = ModMinecraft.McMcInstanceSelected.Name;
+                            args[0] = ModInstanceList.McMcInstanceSelected.Name;
                         }
                         ModBase.RunInUi(() =>
                         {
@@ -320,7 +320,7 @@ namespace PCL
                         if (args.Length == 1)
                             throw new Exception($"EventType {type} 需要至少 2 个以 | 分割的参数，例如 UiLauncherTransparent|400");
                         if (ConfigService.TryGetConfigItemNoType(args[0], out var item) && item.Source != ConfigSource.SharedEncrypt)
-                            item.SetValueNoType(args[1], ModMinecraft.McMcInstanceSelected?.PathInstance);
+                            item.SetValueNoType(args[1], ModInstanceList.McMcInstanceSelected?.PathInstance);
                         if (args.Length == 2)
                             ModMain.Hint($"已写入设置：{args[0]} → {args[1]}", ModMain.HintType.Finish);
                         break;
