+++ Plain Craft Launcher 2/Pages/PageInstance/PageInstanceCompResource.xaml.cs
@@ -1024,7 +1024,7 @@ public partial class PageInstanceCompResource : IRefreshable
         }
 
         // 3. Determine target instance
-        var targetInstance = ModMinecraft.McMcInstanceSelected;
+        var targetInstance = ModInstanceList.McMcInstanceSelected;
         if (ModMain.frmMain.pageCurrent == FormMain.PageType.InstanceSetup) targetInstance = PageInstanceLeft.McInstance;
 
         // 4. Validate instance status
@@ -1121,7 +1121,7 @@ public partial class PageInstanceCompResource : IRefreshable
         }
 
         // 获取并检查目标实例
-        var targetInstance = ModMinecraft.McMcInstanceSelected;
+        var targetInstance = ModInstanceList.McMcInstanceSelected;
         if (ModMain.frmMain.pageCurrent == FormMain.PageType.InstanceSetup)
             targetInstance = PageInstanceLeft.McInstance;
